+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # MarkersExtractor Change Log
 
+## [0.3.16](https://github.com/TheAcharya/MarkersExtractor/releases/tag/0.3.16) (2025-07-??)
+
+### New Features
+
+- Added Apple Compressor export profile to export chapter marker lists (#123)
+
+### Maintenance
+
+- Organized CLI help text arguments into categories (#121)
+
 ## [0.3.15](https://github.com/TheAcharya/MarkersExtractor/releases/tag/0.3.15) (2025-03-29)
 
 ### Bug Fixes